V2 new player (#55)
Anesidora is now at version 2! Thanks to some huge contributions by Hugh the player has been fully redesigned with a new look and feel! Hugh also worked to add the option to theme the player so that you can customize it however you'd like. Of course if you're a fan of the original design, you can continue to use it the same way you always have.

Full update list:

    Player redesign Hugh
    Theming Hugh
    Theme builder Hugh
    Theme import/export Hugh
    Theme repository
    Fixed song history Hugh
    Added ability to Dislike from song History Hugh
    A dynamically updating Options page to preview size changes Hugh
    Several code quality changes to make future updates easier
    A few small tweaks to the Chrome options page due to limitations in Chrome